refactor: Rename `lib/plugins/aws/utils/resolveCfRefValue.js`
