+++ lib/plugins/aws/invoke-local/index.js
@@ -22,7 +22,7 @@ const fileExists = require('../../../utils/fs/file-exists');
 const isStandalone = require('../../../utils/is-standalone-executable');
 const ensureArtifact = require('../../../utils/ensure-artifact');
 const resolveCfImportValue = require('../utils/resolve-cf-import-value');
-const resolveCfRefValue = require('../utils/resolveCfRefValue');
+const resolveCfRefValue = require('../utils/resolve-cf-ref-value');
 const { writeText, progress, log } = require('@serverless/utils/log');
 
 const cachePath = path.join(cachedir('serverless'), 'invokeLocal');
