+++ lib/plugins/aws/utils/resolve-cf-ref-value.js
