+++ test/unit/lib/plugins/aws/utils/resolve-cf-ref-value.test.js
@@ -2,7 +2,7 @@
 
 const expect = require('chai').expect;
 const BbPromise = require('bluebird');
-const resolveCfRefValue = require('../../../../../../lib/plugins/aws/utils/resolveCfRefValue');
+const resolveCfRefValue = require('../../../../../../lib/plugins/aws/utils/resolve-cf-ref-value');
 
 describe('#resolveCfRefValue', () => {
   it('should return matching exported value if found', () => {
